add new util for json and prop file
