+++ core/src/main/java/org/wso2/testgrid/core/phase/InfraPhase.java
@@ -19,13 +19,9 @@
 
 package org.wso2.testgrid.core.phase;
 
-import org.json.JSONObject;
-import org.json.JSONTokener;
-
 import org.wso2.testgrid.common.GrafanaDashboardHandler;
 import org.wso2.testgrid.common.InfrastructureProvider;
 import org.wso2.testgrid.common.InfrastructureProvisionResult;
-import org.wso2.testgrid.common.TestGridConstants;
 import org.wso2.testgrid.common.TestPlanPhase;
 import org.wso2.testgrid.common.TestPlanStatus;
 import org.wso2.testgrid.common.config.InfrastructureConfig;
@@ -38,33 +34,17 @@ import org.wso2.testgrid.common.util.StringUtil;
 import org.wso2.testgrid.core.exception.TestPlanExecutorException;
 import org.wso2.testgrid.infrastructure.InfrastructureProviderFactory;
 
-import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStreamWriter;
-import java.io.PrintWriter;
 
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Properties;
-
-import static java.nio.file.StandardOpenOption.APPEND;
-import static java.nio.file.StandardOpenOption.CREATE;
 
 /**
  * This class includes the implementation of the infrastructure-provisioning phase.
  */
 public class InfraPhase extends Phase {
+
+
 
     @Override
     boolean verifyPrecondition() {
@@ -111,11 +91,17 @@ public class InfraPhase extends Phase {
 
             InfrastructureProvisionResult provisionResult = new InfrastructureProvisionResult();
 
-            persistInfraInputsGeneral();
+            jsonpropFileEditor.persistInfraInputsGeneral(DataBucketsHelper.getInputLocation(getTestPlan())
+                    .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE),
+                    DataBucketsHelper.getInputLocation(getTestPlan())
+                    .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_JSONFILE), getTestPlan());
 
             for (Script script : infrastructureConfig.getFirstProvisioner().getScripts()) {
                 if (!Script.Phase.DESTROY.equals(script.getPhase())) {
-                    persistInfraInputs(script);
+                    jsonpropFileEditor.persistInfraInputs(script, DataBucketsHelper.getInputLocation(getTestPlan())
+                            .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE),
+                            DataBucketsHelper.getInputLocation(getTestPlan())
+                                    .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_JSONFILE));
                     InfrastructureProvider infrastructureProvider = InfrastructureProviderFactory
                             .getInfrastructureProvider(script);
                     infrastructureProvider.init(getTestPlan());
@@ -128,8 +114,12 @@ public class InfraPhase extends Phase {
                         logger.warn("Infra script '" + script.getName() + "' failed. Not running remaining scripts.");
                         break;
                     }
-                    removeScriptParams(script);
-                    refillFromPropFile();
+                    jsonpropFileEditor.removeScriptParams(script, DataBucketsHelper.getInputLocation(getTestPlan())
+                            .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE));
+                    jsonpropFileEditor.refillFromPropFile(DataBucketsHelper.getInputLocation(getTestPlan())
+                            .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE),
+                            DataBucketsHelper.getInputLocation(getTestPlan())
+                            .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_JSONFILE));
                 }
             }
 
@@ -180,303 +170,6 @@ public class InfraPhase extends Phase {
         provisionResult.getProperties().putAll(aProvisionResult.getProperties());
         if (!aProvisionResult.isSuccess()) {
             provisionResult.setSuccess(false);
-        }
-    }
-
-    /**
-     *  Writes Any non script related Infrastructure parameters to the JSON file
-     *  exists as an exterior label
-     *
-     */
-    private void persistInfratoJSON(Map properties, Path jsonFilePath) {
-
-        try {
-            InputStream is = new FileInputStream(jsonFilePath.toString());
-            JSONTokener tokener = new JSONTokener(is);
-            JSONObject inputJson = new JSONObject(tokener);
-            Iterator it = properties.entrySet().iterator();
-            while (it.hasNext()) {
-                Map.Entry pair = (Map.Entry) it.next();
-                if (inputJson.has("general")) {
-                    inputJson.getJSONObject("general").put((String) pair.getKey(), pair.getValue());
-                } else {
-                    JSONObject general = new JSONObject(properties);
-                    inputJson.put("general", general);
-                }
-            }
-
-            try (BufferedWriter writer = Files.newBufferedWriter(Paths.get(jsonFilePath.toString()))) {
-                inputJson.write(writer);
-                writer.write("\n");
-            } catch (Throwable e) {
-                logger.error("Error while persisting infra input params to " + jsonFilePath, e);
-            }
-        } catch (IOException e) {
-            logger.info(jsonFilePath + "created");
-            HashMap<String, Object> insertvalue = new HashMap<String, Object>();
-            insertvalue.put("general", properties);
-            JSONObject inputJson = new JSONObject(insertvalue);
-            try (BufferedWriter jsonWriter = Files.newBufferedWriter(Paths.get(jsonFilePath.toString()))) {
-                inputJson.write(jsonWriter);
-                jsonWriter.write("\n");
-            } catch (Throwable ex) {
-                logger.error("Error while persisting infra input params to " + jsonFilePath, ex);
-            }
-        }
-
-    }
-
-    /**
-     *  Writes Infrastructure input Parameters to the JSON files under the
-     *  proper Script name
-     *
-     */
-    private void persistInfraParamstoJSON(Map properties, Path jsonFilePath, String scriptName) {
-        try {
-            InputStream jsonInputStream = new FileInputStream(jsonFilePath.toString());
-            JSONTokener jsonTokener = new JSONTokener(jsonInputStream);
-            JSONObject inputJson = new JSONObject(jsonTokener);
-            JSONObject general = null;
-            JSONObject scriptParamsJson = new JSONObject();
-            if (inputJson.has("general")) {
-                general = inputJson.getJSONObject("general");
-            }
-            JSONObject scriptParamsOnly = new JSONObject(properties);
-
-            Iterator it = properties.entrySet().iterator();
-            while (it.hasNext()) {
-                Map.Entry pair = (Map.Entry) it.next();
-                scriptParamsJson.put((String) pair.getKey(), pair.getValue());
-            }
-            if (general != null) {
-                Iterator<String> genit = general.keys();
-                while (genit.hasNext()) {
-                    String key = genit.next();
-                    scriptParamsJson.put(key, general.get(key));
-                }
-            }
-            inputJson.put(scriptName, scriptParamsOnly);
-            inputJson.put("currentscript", scriptParamsJson);
-            try (BufferedWriter jsonWriter = Files.newBufferedWriter(Paths.get(jsonFilePath.toString()))) {
-                inputJson.write(jsonWriter);
-                jsonWriter.write("\n");
-            } catch (Throwable e) {
-                logger.error("Error while persisting infra input params to " + jsonFilePath, e);
-            }
-        } catch (IOException e) {
-            logger.info(jsonFilePath + "created");
-            JSONObject json = new JSONObject(properties);
-            JSONObject scriptparams = new JSONObject();
-            scriptparams.put(scriptName, json);
-            scriptparams.put("currentscript", json);
-            try (BufferedWriter writer = Files.newBufferedWriter(Paths.get(jsonFilePath.toString()))) {
-                scriptparams.write(writer);
-                writer.write("\n");
-            } catch (Throwable ex) {
-                logger.error("Error while persisting infra input params to " + jsonFilePath, ex);
-            }
-        }
-    }
-
-    /**
-     * Observes the properties file and updates any necessary values to the json file as a general input
-     *  NOTE :: OVERRIDES ANY PROPERTY OF THE SAME NAME
-
-     */
-    private void refillFromPropFile() {
-
-        final Path propFilePath = DataBucketsHelper.getInputLocation(getTestPlan())
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE);
-        final Path jsonFilePath = DataBucketsHelper.getInputLocation(getTestPlan())
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_JSONFILE);
-
-        InputStream propInputStream = null;
-        InputStream jsonInputStream = null;
-
-        try {
-            propInputStream = new FileInputStream(propFilePath.toString());
-            Properties existingprops = new Properties();
-            existingprops.load(propInputStream);
-            try {
-                jsonInputStream = new FileInputStream(jsonFilePath.toString());
-                JSONTokener jsonTokener = new JSONTokener(jsonInputStream);
-                JSONObject inputJson = new JSONObject(jsonTokener);
-
-                Iterator it = existingprops.entrySet().iterator();
-
-                while (it.hasNext()) {
-                    Map.Entry pair = (Map.Entry) it.next();
-                    if (pair.getValue() != null) {
-                        if (inputJson.has("general")) {
-                            inputJson.getJSONObject("general").put((String) pair.getKey(), pair.getValue());
-                        } else {
-                            JSONObject general = new JSONObject(existingprops);
-                            inputJson.put("general", general);
-                        }
-                    }
-                }
-
-                try (BufferedWriter jsonWriter = Files.newBufferedWriter(Paths.get(jsonFilePath.toString()))) {
-                    inputJson.write(jsonWriter);
-                    jsonWriter.write("\n");
-                } catch (Throwable e) {
-                    logger.error("Error while persisting infra input params to " + jsonFilePath);
-                }
-            } catch (IOException ex) {
-                logger.info("ERROR please view log files to Debug, " + propFilePath + " found but " + jsonFilePath +
-                        " file not found ");
-                HashMap<String, Object> insertvalue = new HashMap<String, Object>();
-                insertvalue.put("general", existingprops);
-                JSONObject inputjson = new JSONObject(insertvalue);
-                try (BufferedWriter jsonWriter = Files.newBufferedWriter(Paths.get(jsonFilePath.toString()))) {
-                    inputjson.write(jsonWriter);
-                    jsonWriter.write("\n");
-                } catch (Throwable exc) {
-                    logger.error("Error while persisting infra input params to " + jsonFilePath);
-                }
-            } finally {
-                try {
-                    if (jsonInputStream != null) {
-                        jsonInputStream.close();
-                    }
-                } catch (Exception e) {
-                    logger.error("Failed to close Stream");
-                }
-            }
-
-        } catch (FileNotFoundException e) {
-            logger.info(propFilePath + " Not created yet ignoring read property file step");
-        } catch (IOException e) {
-            logger.info(propFilePath + " Not created yet ignoring read property file step");
-        } finally {
-            try {
-                if (propInputStream != null) {
-                    propInputStream.close();
-                }
-            } catch (Exception e) {
-                logger.error("Failed to close Stream");
-            }
-        }
-
-    }
-
-    /**
-     *
-     *  Optional method removes a scripts params from prop file after execution
-     */
-
-    private void removeScriptParams(Script script) {
-        final Path propFilePath = DataBucketsHelper.getInputLocation(getTestPlan())
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE);
-        InputStream propInputStream = null;
-        try {
-            propInputStream = new FileInputStream(propFilePath.toString());
-            Properties existingprops = new Properties();
-            existingprops.load(propInputStream);
-
-            Map<String, Object> inputParams = script.getInputParameters();
-
-            Iterator it = inputParams.entrySet().iterator();
-            while (it.hasNext()) {
-                Map.Entry pair = (Map.Entry) it.next();
-                if (existingprops.containsKey(pair.getKey())) {
-                    existingprops.remove(pair.getKey());
-                    logger.info("removing property " + pair.getKey());
-                }
-            }
-
-            try (PrintWriter printWriter = new PrintWriter(
-                    new OutputStreamWriter(Files.newOutputStream(propFilePath), StandardCharsets.UTF_8))) {
-                Iterator itr = existingprops.entrySet().iterator();
-                while (itr.hasNext()) {
-                    Map.Entry pair = (Map.Entry) itr.next();
-                    printWriter.println(pair.getKey() + "=" + pair.getValue());
-                }
-            } catch (IOException e) {
-                logger.error("Error while persisting infra input params to " + propFilePath, e);
-            }
-
-        } catch (FileNotFoundException e) {
-            logger.info(propFilePath + " Not created yet ignoring read property file step");
-        } catch (IOException e) {
-            logger.info(propFilePath + " Failed to read file");
-        } finally {
-            try {
-                if (propInputStream != null) {
-                    propInputStream.close();
-                }
-            } catch (Exception e) {
-                logger.error("Failed to close Stream");
-            }
-        }
-    }
-    /**
-     * The infra-provision.sh / deploy.sh / run-scenario.sh receive the test plan
-     * configuration as a properties file aswell as a json file
-     *
-     * @params script script to be added to databucket files
-     *
-     */
-    private void persistInfraInputs(Script script) {
-
-
-
-        final Path location = DataBucketsHelper.getInputLocation(getTestPlan())
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE);
-        final Path jsonlocation = DataBucketsHelper.getInputLocation(getTestPlan())
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_JSONFILE);
-
-        Map<String, Object> inputParams = script.getInputParameters();
-        persistInfraParamstoJSON(inputParams, jsonlocation, script.getName());
-
-        try (PrintWriter printWriter = new PrintWriter(
-                new OutputStreamWriter(Files.newOutputStream(location, CREATE, APPEND), StandardCharsets.UTF_8))) {
-            Iterator it = inputParams.entrySet().iterator();
-            while (it.hasNext()) {
-                Map.Entry pair = (Map.Entry) it.next();
-                printWriter.println(pair.getKey() + "=" + pair.getValue());
-            }
-        } catch (IOException e) {
-            logger.error("Error while persisting infra input params to " + location, e);
-        }
-    }
-
-    /**
-     *  Write infraParams and job properties to .json file .properties file
-     *
-     */
-    private void persistInfraInputsGeneral() {
-
-        refillFromPropFile();
-
-        final Path location = DataBucketsHelper.getInputLocation(getTestPlan())
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE);
-        final Path jsonlocation = DataBucketsHelper.getInputLocation(getTestPlan())
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_JSONFILE);
-        final Properties infraParameters = getTestPlan().getInfrastructureConfig().getParameters();
-        final Properties jobProperties = getTestPlan().getJobProperties();
-        final String keyFileLocation = getTestPlan().getKeyFileLocation();
-
-        if (keyFileLocation != null) {
-            jobProperties.setProperty(TestGridConstants.KEY_FILE_LOCATION, keyFileLocation);
-        }
-        persistInfratoJSON(jobProperties, jsonlocation);
-        persistInfratoJSON(infraParameters, jsonlocation);
-
-        try (PrintWriter printWriter = new PrintWriter(
-                new OutputStreamWriter(Files.newOutputStream(location, CREATE, APPEND), StandardCharsets.UTF_8))) {
-            for (Enumeration en = jobProperties.propertyNames(); en.hasMoreElements();) {
-                String key = (String) en.nextElement();
-                printWriter.println(key + "=" + jobProperties.getProperty(key));
-            }
-            for (Enumeration en = infraParameters.propertyNames(); en.hasMoreElements();) {
-                String key = (String) en.nextElement();
-                printWriter.println(key + "=" + infraParameters.getProperty(key));
-            }
-
-
-        } catch (IOException e) {
-            logger.error("Error while persisting infra input params to " + location, e);
         }
     }
 }
